Don’t require email address for user
Sending this as a PR as I’m not sure why this requirement made it into Kirby back then.

As far as I can tell, the email is only used for the „Send email“ button in the Panel. We should add a check there and hide it if there is no valid email.